removed an unnecessary navigation file
